// src/types.ts
export interface SalesActivity {
  id: string;
  type: 'DEAL_WON' | 'MEETING' | 'CALL' | 'TASK' | 'EMAIL';
  title: string;
  user: {
    name: string;
    avatar: string;
  };
  relatedCustomer?: {
    id: string;
    name: string;
  };
  details: string;
  timestamp: string;
}
// ...
export interface Deal {
  id: string;
  company: string;
  contactPerson?: string;
  description: string;
  value: number;
  dueDate: string;
  assignees: { name: string; avatar: string }[];
  comments: number;
  attachments: number;
  stage: string;
  priority: 'high' | 'medium' | 'low';
  probability: number;
  daysInStage: number;
  tags?: string[];
}